fix - corrigindo problemas de módulos e adicionando repositories dentro do módulo de repositórios
