+++ .prettierrc.json
@@ -4,7 +4,6 @@
     "trailingComma": "all",
     "printWidth": 80,
     "bracketSpacing": true,
-    "tabWidth": 2,
     "useTabs": true,
     "semi": false,
     "quoteProps": "as-needed"
